test_complete.py
#!/usr/bin/env python3
"""
测试 NomadEcho 完整功能
包括：数据持久化、聊天、海报生成等
"""

import json
from pathlib import Path
import sys
from datetime import datetime
from PIL import Image

def load_chats():
    """加载所有聊天对话"""
    chat_file = Path("chats.json")
    if chat_file.exists():
        with open(chat_file, "r", encoding="utf-8") as f:
            return json.load(f)
    return {}

def save_chats(chats_data):
    """保存聊天对话到文件"""
    with open("chats.json", "w", encoding="utf-8") as f:
        json.dump(chats_data, f, ensure_ascii=False, indent=2)

def get_or_create_conversation(sender_id, recipient_id):
    """获取或创建对话ID"""
    conv_id = "_".join(sorted([sender_id, recipient_id]))
    return conv_id

def add_message(sender_id, recipient_id, text):
    """添加新消息"""
    chats_data = load_chats()
    conv_id = get_or_create_conversation(sender_id, recipient_id)
    
    if conv_id not in chats_data:
        chats_data[conv_id] = {
            "participants": sorted([sender_id, recipient_id]),
            "messages": []
        }
    
    message = {
        "sender": sender_id,
        "text": text,
        "timestamp": datetime.now().strftime("%Y-%m-%d %H:%M:%S")
    }
    chats_data[conv_id]["messages"].append(message)
    save_chats(chats_data)

def get_conversation(sender_id, recipient_id):
    """获取指定对话的所有消息"""
    chats_data = load_chats()
    conv_id = get_or_create_conversation(sender_id, recipient_id)
    
    if conv_id in chats_data:
        return chats_data[conv_id]["messages"]
    return []

def extract_keywords(messages):
    """从对话中提取关键词"""
    import re
    
    if not messages:
        return "soul connection, traveling, sharing"
    
    all_text = " ".join([msg["text"] for msg in messages])
    stop_words = {"的", "是", "在", "了", "和", "有", "为", "这", "我", "你", "他", "什么", "怎么", "哪里", "吗", "呢", "啊"}
    words = re.findall(r'[\u4e00-\u9fa5]{2,}', all_text)
    keywords = [w for w in words if w not in stop_words][:5]
    
    if not keywords:
        keywords = ["soul connection", "traveling", "sharing"]
    
    return ", ".join(keywords[:3])

def generate_placeholder_poster(keywords):
    """生成占位海报图片"""
    from PIL import Image, ImageDraw, ImageFont
    
    width, height = 800, 600
    image = Image.new('RGB', (width, height), color='#f5f5f7')
    draw = ImageDraw.Draw(image, 'RGBA')
    
    # 绘制渐变
    for y in range(height):
        r = int(245 - (y / height) * 30)
        g = int(245 - (y / height) * 50)
        b = int(247 - (y / height) * 10)
        draw.rectangle([(0, y), (width, y+1)], fill=(r, g, b))
    
    # 装饰圆形
    circle_color = (102, 126, 234, 50)
    draw.ellipse([(100, 50), (300, 250)], outline=circle_color, width=2)
    draw.ellipse([(500, 350), (750, 550)], outline=circle_color, width=2)
    
    text_main = "✨ 灵魂回响海报 ✨"
    text_keywords = keywords
    text_subtitle = "这是你们灵魂相遇的瞬间"
    
    try:
        font_main = ImageFont.truetype("/usr/share/fonts/truetype/dejavu/DejaVuSans-Bold.ttf", 48)
        font_sub = ImageFont.truetype("/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf", 28)
        font_small = ImageFont.truetype("/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf", 20)
    except:
        font_main = font_sub = font_small = ImageFont.load_default()
    
    text_y = height // 2 - 60
    draw.text((width//2, text_y), text_main, fill=(102, 126, 234), font=font_main, anchor="mm")
    draw.text((width//2, text_y + 80), text_keywords, fill=(118, 75, 162), font=font_sub, anchor="mm")
    draw.text((width//2, height - 100), text_subtitle, fill=(149, 165, 166), font=font_small, anchor="mm")
    
    return image

def test_chat_functions():
    """测试聊天相关的数据函数"""
    print("=" * 60)
    print("🧪 NomadEcho 聊天功能测试")
    print("=" * 60)
    
    # 测试 1: 创建对话 ID
    print("\n✅ 测试 1: 创建对话 ID")
    conv_id = get_or_create_conversation("user_123", "user_456")
    print(f"   对话 ID: {conv_id}")
    assert "_" in conv_id, "对话 ID 应该包含下划线"
    print("   ✓ 对话 ID 格式正确\n")
    
    # 测试 2: 添加消息
    print("✅ 测试 2: 添加聊天消息")
    add_message("user_123", "user_456", "你好！在高山之巅看日出，感到生命的意义")
    add_message("user_456", "user_123", "太美了！我也曾在雪山顶端思考过人生")
    add_message("user_123", "user_456", "我们都在旅行中找到了答案")
    print("   ✓ 消息添加成功\n")
    
    # 测试 3: 读取对话
    print("✅ 测试 3: 读取聊天历史")
    messages = get_conversation("user_123", "user_456")
    print(f"   获取消息数: {len(messages)}")
    for i, msg in enumerate(messages, 1):
        preview = msg['text'][:40] + "..." if len(msg['text']) > 40 else msg['text']
        print(f"   消息 {i}: [{msg['sender']}] {preview}")
    assert len(messages) >= 3, "应该至少有 3 条消息"
    print("   ✓ 消息读取成功\n")
    
    # 测试 4: 验证 JSON 文件
    print("✅ 测试 4: 验证 chats.json 文件")
    chats_file = Path("chats.json")
    if chats_file.exists():
        with open(chats_file, "r", encoding="utf-8") as f:
            data = json.load(f)
        print(f"   文件大小: {chats_file.stat().st_size} 字节")
        print(f"   对话数: {len(data)}")
        print("   ✓ 文件存储成功\n")
    else:
        print("   ⚠️ chats.json 文件不存在\n")

def test_poster_functions():
    """测试海报生成功能"""
    print("=" * 60)
    print("🎨 NomadEcho 海报生成测试")
    print("=" * 60)
    
    # 获取对话
    messages = get_conversation("user_123", "user_456")
    print(f"\n✅ 测试 1: 关键词提取")
    keywords = extract_keywords(messages)
    print(f"   提取的关键词: {keywords}")
    assert keywords, "应该能提取关键词"
    print("   ✓ 关键词提取成功\n")
    
    # 生成占位海报
    print("✅ 测试 2: 生成占位海报")
    poster = generate_placeholder_poster(keywords)
    print(f"   海报尺寸: {poster.size}")
    print(f"   海报模式: {poster.mode}")
    assert poster.size == (800, 600), "海报尺寸应该是 800x600"
    
    # 保存海报用于检查
    poster_path = Path("test_poster.png")
    poster.save(poster_path)
    print(f"   已保存: {poster_path}")
    print(f"   文件大小: {poster_path.stat().st_size} 字节")
    print("   ✓ 海报生成成功\n")
    
    # 清理
    poster_path.unlink()

def test_complete_workflow():
    """测试完整工作流"""
    print("=" * 60)
    print("🔄 NomadEcho 完整工作流测试")
    print("=" * 60)
    
    print("\n✅ 工作流：分享感悟 → 聊天 → 生成海报")
    print("   1️⃣ 分享感悟：✓ 已在 insights_data.json 中实现")
    print("   2️⃣ 发起聊天：✓ 已测试")
    messages = get_conversation("user_123", "user_456")
    print(f"   3️⃣ 生成海报：✓ 关键词 = {extract_keywords(messages)}")
    print("   4️⃣ 展示标题：✓ '这是你们灵魂相遇的瞬间'\n")

def cleanup():
    """清理测试文件"""
    print("🧹 清理测试数据...")
    chats_file = Path("chats.json")
    if chats_file.exists():
        chats_file.unlink()
        print("   已删除 chats.json")
    print()

if __name__ == "__main__":
    try:
        test_chat_functions()
        test_poster_functions()
        test_complete_workflow()
        cleanup()
        
        print("=" * 60)
        print("✨ 所有测试通过！应用已准备就绪")
        print("=" * 60)
        print("\n🚀 启动应用:")
        print("   streamlit run app.py\n")
        
    except AssertionError as e:
        print(f"\n❌ 测试失败: {e}")
        sys.exit(1)
    except Exception as e:
        print(f"\n❌ 错误: {e}")
        import traceback
        traceback.print_exc()
        sys.exit(1)
