requirements.txt
streamlit==1.28.1
pandas==2.1.3
streamlit-autorefresh==0.0.1
requests==2.31.0
pillow==10.1.0
python-dotenv==1.0.0
